+++ CHANGELOG.md
@@ -20,6 +20,8 @@ The current file describes significant changes in the code that may affect the u
 * Added theme `configuration` options into `\Oro\Bundle\FrontendBundle\Resources\views\layouts\default\theme.yml`.
 * Added `oro_theme.theme_configuration` system configuration option.
 * Added `\Oro\Bundle\FrontendBundle\Layout\Extension\PageTemplatesThemeConfigurationExtension` that gets `product_page_template` theme configuration option values and adds for page_templates.
+* Added `\Oro\Bundle\FrontendBundle\Form\Configuration\AbstractCssConfigBuilder` that should be used for CSS Theme Variables
+* Added option `show_input_control` to `\Oro\Bundle\FormBundle\Form\Type\OroSimpleColorPickerType` that allows to show text input form for manual color value. By default, value for this option is false
 
 ### Changed
 * Replaced all places in code that used old system configuration options on theme configuration options.